July 7 update
Added CreatePasswordInAppView, to create password for encryption.
Edit the HomePageView so that it checks if password is not made then go to create password, if made then go to unlocked or locked view. Now newly created password is stored in keychain and retrieved when going to be used for encryption or decryption.